Allow installed repository to not exist in update
Otherwise we have a chicken/egg problem here.

+++ src/Command/InstalledRepositoryLoadingCommandTrait.php
@@ -7,13 +7,17 @@ namespace Phpcq\Command;
 use Phpcq\Exception\RuntimeException;
 use Phpcq\Platform\PlatformInformation;
 use Phpcq\Repository\InstalledRepositoryLoader;
+use Phpcq\Repository\Repository;
 use Phpcq\Repository\RepositoryInterface;
 
 trait InstalledRepositoryLoadingCommandTrait
 {
-    private function getInstalledRepository(string $phpcqPath): RepositoryInterface
+    private function getInstalledRepository(string $phpcqPath, bool $failIfNotExist = true): RepositoryInterface
     {
         if (!is_file($phpcqPath . '/installed.json')) {
+            if (!$failIfNotExist) {
+                return new Repository(PlatformInformation::createFromCurrentPlatform());
+            }
             throw new RuntimeException('Please install the tools first ("phpcq update").');
         }
         $loader = new InstalledRepositoryLoader(PlatformInformation::createFromCurrentPlatform());
